% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Title}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Angle measured in degrees}
}
\value{
Angle measured in radians
}
\description{
Title
}
\examples{
deg2rad(90)
deg2rad(285)
}
