% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_distance.R
\name{great_circle_distance}
\alias{great_circle_distance}
\title{Title}
\usage{
great_circle_distance(lat1, lat2, lon1, lon2, scale = 6371)
}
\arguments{
\item{lat1}{The latitude of the first location}

\item{lat2}{The latitude of the second location}

\item{lon1}{The longitude of the first location}

\item{lon2}{The longitude of the second location}

\item{scale}{The radius of the sphere in kilometers}
}
\value{
The great circle distance between the first location and the second location
}
\description{
Title
}
\examples{
great_circle_distance(-34, 18, 41, -74)

}
